fixed bugs @PCA

+++ include/pca.h
@@ -4,6 +4,6 @@
 #include <gsl/gsl_matrix.h>
 #include "OPF.h"
 
-Subgraph *PCA(Subgraph *in, int p); /* It performs the standard Principal Component Analisys algorithm */
+Subgraph *PCA(Subgraph *in, double p); /* It performs the standard Principal Component Analisys algorithm */
 
 #endif
